requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.0
openpyxl==3.1.2
Werkzeug==2.3.7
Pillow==10.0.0